ADD: basic architecture according to feature-sliced.design
