+++ src/app/providers/ThemeProviders/index.tsx
@@ -1,0 +1,4 @@
+import { UseTheme } from "./lib/UseTheme";
+import { ThemeProvider } from "./ui/ThemeProvider";
+
+export { ThemeProvider, UseTheme };
