+++ src/app/providers/ThemeProviders/lib/UseTheme.tsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from "./ThemeContext";
+import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from "app/providers/ThemeProviders/lib/ThemeContext";
 
 interface UseThemeResult {
   toggleTheme: () => void;
