+++ src/app/styles/index.scss
@@ -1,0 +1,13 @@
+@import 'reset';
+@import 'themes/dark.scss';
+@import 'themes/normal.scss';
+@import 'variables/global.scss';
+
+.app {
+    font: var(--font-l);
+
+    background: var(--bg-color);
+    color: var(--primary-color);
+
+    min-height: 100vh;
+}
