--- src/components/Counter.module.scss
@@ -1,7 +1,0 @@
-.incr-button {
-    background-color: indigo;
-}
-
-.btnGreen {
-    background-color: aquamarine;
-}
