+++ src/index.tsx
@@ -1,13 +1,11 @@
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import { App } from "./App";
-import { ThemeProvider } from "./theme/ThemeProvider";
+import { App } from "./app/App";
+import { ThemeProvider } from "app/providers/ThemeProviders";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
-  // BrowserRouter - обертка для того, чтобы работал роутинг
   <BrowserRouter>
-    {/* Обертка контекста темы */}
     <ThemeProvider>
       <App />
     </ThemeProvider>
