+++ src/pages/AboutPage/index.tsx
@@ -1,0 +1,18 @@
+// согласно применяемой методологии, для каждого слоя должны быть сегменты и слайсы
+// (за исключением shared), а также Public api, которым будет выступать файл
+// index.tsx
+
+// поэтому закинем файлы в папку ui а в корне сделаем index.tsx
+// Наружу выкенем асинхронный чанк, обычный компонент выкидывать не будем
+
+// При этом переименуем экспорт для красоты
+
+// Стоит отметить, что автоматически редактор предложит сделать так:
+        // import { AboutPageAsync } from "./ui/AboutPage.async";
+        // export { AboutPageAsync as AboutPage };
+// Однако импорт можно опустить, если сразу же написать откуда берем то что экспортируем:
+
+export { AboutPageAsync as AboutPage } from "./ui/AboutPage.async"
+
+// Внтури модуля пути стоит делать отрносительными т.к. при его переносе куда либо
+// благодаря этому не нужно будет менять пути
