+++ src/pages/AboutPage/ui/AboutPage.tsx
