+++ src/pages/MainPage/index.tsx
@@ -1,0 +1,1 @@
+export {MainPageAsync as MainPage} from './ui/MainPage.async'
