--- src/styles/index.scss
@@ -1,18 +1,0 @@
-@import 'reset';
-@import 'themes/dark.scss';
-@import 'themes/normal.scss';
-@import 'variables/global.scss';
-
-// Это файл с глобальными стилями. Однако с учетом сделанных настроек ВебПака он тоже
-// инкапсулирован и его нужно сделать глобальным. Это сделано в конфиге сборщика
-
-
-.app {
-    // подключаем шрифты по умолчанию
-    font: var(--font-l);
-
-    background: var(--bg-color);
-    color: var(--primary-color);
-
-    min-height: 100vh;
-}
