+++ tsconfig.json
@@ -1,24 +1,21 @@
 {
     "compilerOptions": {
-      // Куда производится сборка
       "outDir": "./dist/",
-      // Подсвечивает все места где не указан тип 
       "noImplicitAny": true,
-      // указывается используемая модульная система
       "module": "ESNext",
-      // в какую спецификацию компилируется
       "target": "es5",
-      // Позволяет использовать jsx, как в Реакт
       "jsx": "react-jsx",
-      // Позволяет компилятору обрабытывать js файлы
       "allowJs": true,
-      // Определяет какие импорты будут использоваться 
       "moduleResolution": "node",
-      // тоже для импортов. Позволяет работать с пакетами, которые используют common js
       "esModuleInterop": true,
-      // Используется для более удобной работы с импортами (чтобы не использовать * as...)
       "allowSyntheticDefaultImports": true,
       "baseUrl": ".",
+      // Эта запись говорит о том, что все абсолютные экспорт и импорты
+      // будут производиться из папки src
+      // Чтобы это работало нужно также настроить и вебпак
+      "paths": {
+        "*": ["./src/*"]
+      },
     },
     "ts-node": {
       "compilerOptions": {
