+++ webpack.config.ts
@@ -8,6 +8,8 @@ export default (env: BuildEnv) => {
     entry: path.resolve(__dirname, "src", "index.tsx"),
     build: path.resolve(__dirname, "bundle"),
     html: path.resolve(__dirname, "public", "index.html"),
+    // путь до папки с исходным кодом
+    src: path.resolve(__dirname, "src"),
   };
 
   const mode = env.mode || "development";
